Handles user lists, asks for user lists on plugin loads, and more
PieSpy will now ask for a user list on __init__. It will then handle receiving
a list of users, and will feed them into the PieSpy database. Now, someone
doesn't have to talk first before they can be connected to someone else on
the graph - all they have to do is be in the channel.

Bugfix: doLayout() in pyPie was fixed so that it only will operate on the
nodes that will be visible on the graph, which will cut the time it takes
that function to run by a whole bunch. This bug made itself apparent once I
had implemented the user list check and tried graphing a ~25 user channel.

I also changed things so that it will squeeze the graph into a nicer area, so
nodes won't be all over those three lines of text on the bottom.